+++ vital/kpf/examples/kpf_example.cxx
@@ -10,6 +10,7 @@
 #include <vital/kpf/kpf_parse.h>
 #include <vital/kpf/kpf_text_parser.h>
 #include <vital/kpf/kpf_canonical_io_adapter.h>
+#include <vital/kpf/kpf_text_writer.h>
 #include <iostream>
 #include <sstream>
 #include <string>
