+++ vital/kpf/examples/kpf_example_simple.cxx
@@ -10,6 +10,8 @@
 
 #include <vital/kpf/kpf_parse.h>
 #include <vital/kpf/kpf_text_parser.h>
+#include <vital/kpf/kpf_text_writer.h>
+
 #include <iostream>
 #include <sstream>
 #include <string>
